requirements.txt
bson==0.5.8
certifi==2020.12.5
dataclasses==0.6
Django==2.2.12
djangorestframework==3.11.0
djongo==1.3.0
mccabe==0.6.1
pymongo==3.7.2
python-dateutil==2.8.1
rest-meets-djongo==0.0.13
sqlparse==0.2.4
dnspython==2.1.0
django-cors-headers==3.6.0
requests==2.25.1
beautifulsoup4==4.9.3
django-google-api==1.0.1